% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_study.R
\name{extract_genes}
\alias{extract_genes}
\title{Extract the master gene symbol vector from an `ExpressionSet`.}
\usage{
extract_genes(expression_eset, gene_symbol_col = "geneSymbol")
}
\arguments{
\item{expression_eset}{A Biobase `ExpressionSet`.}

\item{gene_symbol_col}{Column name in `fData(expression_eset)`.}
}
\value{
Character vector of gene symbols (length = number of rows).
}
\description{
Extract the master gene symbol vector from an `ExpressionSet`.
}
